Adjust serial comms, build

+++ source/Attenuator/Header.h
@@ -189,19 +189,19 @@ enum MENU_LEVELS MENU_LEVEL;
 SerialTransfer packComs;
 
 struct __attribute__((packed)) STRUCT {
-  int s;
-  int i;
-  int d1;
-  int d2;
-  int e;
+  uint16_t s;
+  uint16_t i;
+  uint16_t d1;
+  uint16_t d2;
+  uint16_t e;
 } comStruct;
 
 struct __attribute__((packed)) STRUCTSEND {
-  int s;
-  int i;
-  int d1;
-  int d2;
-  int e;
+  uint16_t s;
+  uint16_t i;
+  uint16_t d1;
+  uint16_t d2;
+  uint16_t e;
 } sendStruct;
 
 /*
